requirements.txt
alembic==1.0.10
asn1crypto==0.24.0
autopep8==1.4.4
bootstrap4==0.1.0
certifi==2019.3.9
cffi==1.12.3
chardet==3.0.4
Click==7.0
cryptography==2.6.1
dominate==2.3.5
electron==1.0.0
Flask==1.0.2
Flask-Bootstrap4==4.0.2
Flask-Migrate==2.4.0
Flask-SQLAlchemy==2.4.0
Flask-WTF==0.14.2
gunicorn==19.9.0
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.1
Mako==1.0.9
MarkupSafe==1.1.1
ndg-httpsclient==0.5.1
nose==1.3.7
psycopg2==2.8.2
pusher==2.1.3
pyasn1==0.4.5
pycodestyle==2.5.0
pycparser==2.19
PyNaCl==1.3.0
pyOpenSSL==19.0.0
python-dateutil==2.8.0
python-editor==1.0.4
python-engineio==3.5.2
python-socketio==3.1.2
requests==2.21.0
six==1.12.0
SQLAlchemy==1.3.3
urllib3==1.24.2
virtualenv==16.5.0
visitor==0.1.3
Werkzeug==0.15.2
WTForms==2.2.1
